refactor(test): move TestListEvents into sdktest/events/ subfolder
Correct the previous misdirected refactor — TestListEvents.java belongs
in sdktest/events/ (matching the reports/ pattern), not EventTest.java.
Also restores EventTest.java to com.smartsheet.api.models.
