+++ src/test/java/com/smartsheet/api/models/EventTest.java
@@ -14,10 +14,9 @@
  * limitations under the License.
  */
 
-package com.smartsheet.api.models.events;
+package com.smartsheet.api.models;
 
 import com.smartsheet.api.internal.json.JacksonJsonSerializer;
-import com.smartsheet.api.models.Event;
 import com.smartsheet.api.models.enums.EventObjectType;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
@@ -27,7 +26,7 @@ import java.nio.charset.StandardCharsets;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class TestListEventsModel {
+class EventTest {
 
     private static Event deserialize(String json) throws Exception {
         JacksonJsonSerializer serializer = new JacksonJsonSerializer();
